Separate user cache into its own file
